+++ .gitignore
@@ -4,7 +4,6 @@
 .helix
 [Bb]inlog/
 [Bb][Uu][Ii][Ll][Dd]/
-[Cc][Mm][Aa][Kk][Ee]/
 [Oo][Bb][Jj]/
 [Oo][Uu][Tt]/
 [Tt][Mm][Pp]/
