+++ launch/dynamo-run/Cargo.toml
@@ -28,12 +28,12 @@ description = "Dynamo Run CLI"
 # Build with `--no-default-features` to disable these defaults
 # We don't include llamacpp by default until we figure out when it needs external libraries
 default = ["mistralrs", "vllm", "sglang"]
-trtllm = ["dynamo-llm/trtllm"]
 python = ["dynamo-llm/python"]
 mistralrs = ["dep:dynamo-engine-mistralrs"]
 llamacpp = ["dep:dynamo-engine-llamacpp"]
 vllm = ["dep:dynamo-engine-vllm", "dep:netlink-packet-route", "dep:rtnetlink"]
 sglang = ["dep:dynamo-engine-sglang", "dep:netlink-packet-route", "dep:rtnetlink"]
+trtllm = ["dep:dynamo-engine-trtllm"]
 
 cuda = ["dynamo-engine-llamacpp/cuda", "dynamo-engine-mistralrs/cuda"]
 metal = ["dynamo-engine-llamacpp/metal", "dynamo-engine-mistralrs/metal"]
@@ -47,6 +47,7 @@ dynamo-engine-llamacpp = { path = "../../lib/engines/llamacpp", optional = true 
 dynamo-engine-mistralrs = { path = "../../lib/engines/mistralrs", optional = true }
 dynamo-engine-sglang = { path = "../../lib/engines/sglang", optional = true }
 dynamo-engine-vllm = { path = "../../lib/engines/vllm", optional = true }
+dynamo-engine-trtllm = { path = "../../lib/engines/trtllm", optional = true }
 
 anyhow = { workspace = true }
 async-stream = { workspace = true }
