+++ lib/engines/trtllm/src/executor/protocols/stats.rs
